third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-datachannel-expected.txt

Tests RTCDataChannel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dc = pc.createDataChannel("label1"); did not throw exception.
PASS dc.reliable is true
PASS dc = pc.createDataChannel("label2", {}); did not throw exception.
PASS dc.reliable is true
PASS dc = pc.createDataChannel("label3", {ordered:true}); did not throw exception.
PASS dc.reliable is true
PASS dc = pc.createDataChannel("label3", {ordered:false}); did not throw exception.
PASS dc.reliable is true
PASS dc.ordered is false
PASS dc = pc.createDataChannel("label3", {maxRetransmits:0}); did not throw exception.
PASS dc.reliable is false
PASS dc = pc.createDataChannel("label3", {maxPacketLifeTime:0}); did not throw exception.
PASS dc.reliable is false
PASS dc = pc.createDataChannel("label3", {maxRetransmitTime:0}); did not throw exception.
PASS dc.reliable is true
PASS dc = pc.createDataChannel("label"); did not throw exception.
PASS dc.readyState is 'connecting'
PASS pc is connected
PASS dc_onopen was called
PASS dc.readyState is 'open'
PASS dc.label is 'label'
PASS pc_ondatachannel was called
PASS dc2.label is 'label'
PASS dc.send('xyzzy'); did not throw exception.
PASS dc_onmessage_string was called
PASS data is 'xyzzy'
PASS dc.send(buffer); did not throw exception.
PASS dc_onmessage_arraybuffer was called
PASS data.byteLength is 2
PASS array[0] is 17
PASS array[1] is 19
PASS data.byteLength is 12
PASS dc.send(shrunkView); did not throw exception.
PASS dc_onmessage_dataview was called
PASS data.byteLength is 10
PASS array[0] is 1
PASS array[9] is 10
PASS dc.send(new Uint8Array(new SharedArrayBuffer(16))); threw exception TypeError: Failed to execute 'send' on 'RTCDataChannel': The provided ArrayBufferView value must not be shared..
PASS dc_onclose was called
PASS dc.readyState is 'closed'
PASS successfullyParsed is true

TEST COMPLETE
